Move name bits to name file
